Handler update and restructure
- refactored folder structure
- abstracted gx logic to fxmodules
