+++ http/todoHandler.go
@@ -1,4 +1,4 @@
-package handlers
+package http
 
 import (
 	"net/http"
